Deleted MODIS AOD Downloader Script
This was the original MODIS AOD Downloader Script used to download the entire world's worth of data from 2008-2014. We are now using the FTP pull method from the EarthData website, as it is better for our study area since the MOD04_L2 dataset does not include geographic information in the filename.